requirements.txt
flask
torch
transformers
PyPDF2
python-docx
werkzeug
flask-cors
pdf2image
pytesseract


